Add user and user profile management
Through the 'accounts' application, we provide a backend capable of managing user accounts registration and activation, user profile editing and user's password updating.

An user can be created by a POST request to the 'user-list'. The creation of a user automatically triggers the creation of his corresponding profile (which right now contains information such as: gender, birth date, location and nationality,and is easily upgradable in the future with new fields), which can be updated with PUT or PATCH requests to 'user-update-profile' endpoint. Furthermore at the end of the registration process, an activation mail will be sent to the user, providing a link that accepts a GET request, which if correctly delivered will set to "True" the is_active flag of the corresponding user object.
Finally a user can change is own password instantiating  a POST request to the 'user-change-password' endpoint.
