+++ accounts/viewsets.py
@@ -8,6 +8,7 @@ from .serializers import UserSerializer, UserPasswordSerializer
 from .permissions import IsAdminOrSelf
 from .models import UserProfile
 from .serializers import UserProfileSerializer
+from . import utils
 
 
 class UserViewSet(viewsets.ModelViewSet):
@@ -44,7 +45,6 @@ class UserViewSet(viewsets.ModelViewSet):
 
         if serializer.is_valid():
             user.set_password(serializer.validated_data['password'])
-
             user.save()
             return Response(status=status.HTTP_200_OK)
         else:
@@ -65,3 +65,36 @@ class UserViewSet(viewsets.ModelViewSet):
         else:
             return Response(serializer.errors,
                             status=status.HTTP_400_BAD_REQUEST)
+
+    @detail_route(methods=['get'], permission_classes=[IsAdminOrSelf])
+    def activate(self, request, upk=None, token=None):
+        # First, if client is accessing to the activation url's root we have to refuse his request
+        if upk is None or token is None:
+            return Response({'msg': 'This activation link is not valid'}, status=status.HTTP_403_FORBIDDEN)
+        # Then we start validating those two fields
+        try:
+            user = self.get_user_to_activate(upk)
+        except get_user_model().DoesNotExist:
+            return Response({'msg': 'This activation link is not valid'}, status=status.HTTP_400_BAD_REQUEST)
+
+        # Now validate the token
+        if utils.validate_token(user, token):  # If it's valid we can start activation process
+            return self.activate_user(user)
+
+        # Otherwise we send a 400
+        return Response({},
+                        status=status.HTTP_400_BAD_REQUEST)
+
+    def activate_user(self, user):
+        if user.is_active:  # If the user is using an activation link being already active we send a 403
+            return Response({},
+                            status=status.HTTP_403_FORBIDDEN)
+        else:  # Otherwise we can activate the account
+            user.is_active = True
+            user.save()
+            return Response({'msg': 'The account had been activated correctly'},
+                            status=status.HTTP_204_NO_CONTENT)
+
+    def get_user_to_activate(self, upk):
+        user = get_user_model().objects.get(pk=upk)
+        return user
